Extract FnSymbol.{h,cpp} from Symbol.{h,cpp}
